Implemented iterative pre-order traversal
We will need this for several operations. Although we did make a
recursive implementation first, the fear of `StackOverflowErrors`
pushed us to use this approach instead. In essence, the method
executes a callback whenever a node is visited for the first time.
The callback should be a void consumer, meaning that although it
will take in the node as argument, it will not return anything.