`jni_md.h` from OpenJDK now included for `build.py`
We include this header file to allow for cross-platform compilation.
The file was pulled from: https://github.com/AdoptOpenJDK/openjdk-jdk11
Since it's listed under `unix` it should work for both Linux and macOS
out of the box. I will also include the one for Windows at some point
when we extend compilation to support it.