Added exceptional behaviour for both `parse` variants
Parsing can return a null pointer in one of three conditions:
1) No or bad language specified for the parser. We protect against this
   at the constructor level;
2) If the cancellation flag is set on the parser, which is currently not
   supported, since we treat parsing as a blocking operation;
3) If the parsing limit duration is exceeded, which is currently the
   only possible failure case.

Considering the aforementioned cases, we can immediately imply that
parsing can realistically only fail in case 3). As such, we accommodate
the failure logic to only consider that case. It would be nice if we
could further expand our parser in the future to support explicit
cancellation, but this is on the back-burner for now.