Added language submodule: Markdown
Note: I had to change build.py to make this work.
For some reason, the owner of that repo makes a distinction
between regular and "inline" markdown. I guess the latest version is ok?
Either way we can revert to the old one if it ends up being unusable.
Shame that this has to "ruin" the homogeneity of the build process.