Added instance method to `Tree` for mapping a `Node` back to content
The visibility of the overloaded `getSource` is currently package
private. We have two possible courses of action now:

1) Retain this level of visibility, and call the method only from `Node`
2) Promote it to `public`, and make it a generally accessible API

The former approach is preferred, but will require a greater degree
of refactoring, since we will have to introduce a reference to each
`Tree` object from its child nodes. Not only that, but all the instance
methods of `Node` used for traversal will also have to copy the tree.
The latter is a no-brainer, and will most likely not require any
additional changes.