The `isExecuted` method of `QueryCursor` is relegated to native code
The reasoning for this is complicated: When declaring a field as final,
the Java compiler automatically replaces occurrences of the field with
the actual value. The finality of `executed` was intended for the Java
end only, while it is still modified on the native side. For this reason
the compiled code will always return a false to calls of `isExecuted`,
regardless of it being modified at the native layer. The best way to
circumvent this is by making the getter itself native, as the Java
compiler will not be able to modify the underlying code. In reality,
the field should not have been declared final in the first place, but
my OCD and autism prevents this. Actually, come to think of it, I doubt
anyone will even use the method. It's convenient for testing purposes
though, but since we already have it, might as well make it public API.