`Parser` instances now created with via builder pattern
On the surface level this might seem like a downgrade, as we were
previously able to construct `Parser` instances directly, and not
via a verbose builder.

The truth is, while that iteration was more straightforward to use,
it was extremely confusing on the implementation side. Since JNI
does not offer support for native constructors, we had to essentially
nest and combine many static methods. I think that the current iteration
is superior in terms of readability and understandability, even though
the users will have to employ more steps to construct a `Parser`.