`Query` instances now created via builder pattern
Unlike adding a builder to `Parser` in 14691e0, `Query` is a far more
fitting candidate as it has more than one field, and thus more
pre-conditions. As was the case with `Parser`, the constructor will be
deprecated and removed in a later version. However, unlike `Parser`, I
think this builder (and the class itself) will be further iterated upon.