Added `TreeCursor#Stub` inner class
It has a similar purpose to `Node#Null`, however due to `TreeCursor`
being an `External`, it needs to have a pointer associated with it.
To emulate a stub, we make use of the `nullptr`, and override all the
instance methods to throw an `UnsupportedOperationException`. In this
way, all classes extending the `Stub` will have to provide their own
implementations of the instance methods. The primary candidate for
extending this inner class is `OffsetTreeCursor`