Improve `TreeSitter` generation
Rather than copying it to the default source directory, we instead
copy it to the `target/generated-sources` directory, and then use a
build helper plugin to include said directory in the compilation and
subsequent sources packaging. This approach has the added benefit of
completely circumventing potential VCS inconsistencies, where generated
files are not committed.