Add script for generating a properties file containing `Language` info
This script is intended to trigger at build time, generating a file
called `language.properties`. Each file consists of a repetition of
three-line sections in the format:

```
tree-sitter.language.{language-name}.url={language-url}
tree-sitter.language.{language-name}.sha={language-sha}
tree-sitter.language.{language-name}.tag={language-tag}
```

Where `language-url` and `language-sha` are the remote URL and full last
commit SHA of the specified language submodule. The `language-tag` is
the current Git tag at which the submodule is situated, and it may or
may not be present. This information will be used at runtime to inject
additional information into `Language` enum values.