Update .gitignore
We want to ignore the generated language metadata properties file.
I've recently come to the realization that generated files should never
see a VCS, and this one is no exception. We only have to take care to
bind the generation to the correct build step, in order to ensure the
file is always picked up for packaging.