Add `Node#getGrammarSymbol`
This also includes a change to the existing `Node#getSymbol`. Both are
now instance methods of `Node`, calling a newly introduced internal
native method `Node#getSymbol(boolean)`. That native method re-uses the
implementation of the old `Node#getSymbol` native, while using the
boolean flag as a switch to determine which `tree-sitter` API to call.