Replace native `TSRange` comparison with an equality function
Comparing ranges did not make sense the way we were doing it initially.
It's far better to do this on the Java end, and just assume that ranges
are properly ordered when they are passed into the native code. The
functionality of `__compareRanges` was stripped down into a dedicated
`__pointEqual` and `__rangeEqual`, which are fare more practical and
make more sense. I'm not sure if these native implementations should
be used from the Java code, especially since its such simple code. We
do implement `Point#compareTo` natively, but even in that case I am
debating whether this is a good idea. At the end of the day, the
`RANGE_FULL` variable is now `RANGE_DEFAULT`, and the method to check
if a `TSRange` matches it is now aptly named `__isDefaultRange`.