Add `Parser#setIncludedRanges`
Added native implementation for a `Parser` mutation method that, as the
name implies, will set the list of ranges that are considered when
parsing documents. Note that while documented, the methods do not yet
check if the range precondition is satisfied or not. That will be
introduced in one of the subsequent commits.