Fix: `Point#compareTo` is no longer native
The method did not work as intended with points containing negative
coordinates. This is because on the JNI end, it was being unmarshalled
into a `TSPoint`, whose coordinates are all defined as _unsigned_ ints.
The `__comparePoints` will remain, as it seems to work fine with the
native methods that use it.