Add new native methods to `TreeSitter` for ABI versioning information
Since we have re-introduced `TreeSitter` as a metaclass, I felt that the
ABI versioning information feels more at place there than it does in the
`Parser`. Conveniently, the generation of this versioning file takes
place right before compilation and the header files can thus be
generated properly on `package`. Subsequent commits will focus on laying
the foundation for sunsetting the corresponding `Parser` methods.