Load `Logger` and `MarkerFactory` classes of SLF4J on JNI startup
Both `jclass` global references will be `NULL` if the dependency classes
are not present. This makes them safe to load globally. Although loading
`Logger` is a no-brainer, the reason we also need a `MarkerFactory` is
because we want to distinguish the `TSLogType` values. Each value,
`TSLogTypeLex` and `TSLogTypeParse` will have their respective markers.