Add `__log_in_java`: native function for logging parser events
The function signature aligns with what is required by `TSLogger.log`.
It first checks if the `payload` is defined, which in our case is an
instance of a `Logger` class. We then obtain the current JVM environment
while attaching the current thread if necessary. We then construct an
instance of `Marker` through the `MarkerFactory#getMarker`, its value
based on the `TSLogType`. Finally, we call the `Logger#debug` method,
passing the constructed `Marker` and a `String` from the character
pointer. Although I'm quite happy with this implementation, we do
restrict all logging to the `DEBUG` level.