Add `startOffset` field to `Pattern`
We were already extracting this information when building out `Query`
instances, but for some reason, we did not include it in the `Pattern`
objects that we constructed. This is no longer the case, and the same
objects now clearly state the starting offset with respect to the
`Query` they reside in.