Add corrected calculation for `getEndOffset`
It's not possible to set the field value because we copy all pattern
characters (extra trailing whitespaces included) into the `Pattern` that
we are constructing. To get around this, we can simply `trim` the value
passed from the JNI when constructing the object. Then we can use the
total string length of the value to calculate the end offset.